requirements.txt
asgiref
Django
sqlparse
tzdata
djangorestframework
djangorestframework-simplejwt
pyjwt
drf-yasg
celery
redis
psycopg2
requests
python-dotenv
